requirements.txt
# Multi-Agent Loan Processor - Dependencies
# Python 3.11+ required

# LLM & Agent Framework
langchain>=0.1.0
langgraph>=0.0.20
langchain-anthropic>=0.1.0
langchain-ollama>=1.0.0
langchain-groq>=0.1.0
anthropic>=0.18.0

# PDF Processing
pymupdf>=1.23.0
pdfplumber>=0.10.0

# UI
streamlit>=1.31.0

# Data & Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Optional: Vector Store for RAG
chromadb>=0.4.0

# Optional: Tracing
langsmith>=0.0.87

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
